Update main.f90
Exploratory changes for #1

+++ src/main.f90
@@ -17,7 +17,7 @@ program GenerateIndicators
     real, dimension(16384)  :: rvPrimarySet
     real, dimension(16384)  :: rvSecondarySet
     type(CompareType)       :: tCmp
-    real                    :: rFB
+    real(8)                 :: rFB
     
     ! First case: Single peak of growing size
     rvPrimarySet   = 1.
